Fixed bug in word generation logic
Problem: The checkIfAllWordsGenerated() function limited the word count to 4 and regenerated the grid when the user got down to 4 words from a higher number. It also prevented actually placing more than 4 words in the grid, despite them showing up in the bank.

Solution: Removed the size check from the checkIfAllWordsGenerated() function and removed an unnecessary call during cell selection.