Major changes - created WordGrid.java & auto-highlighting
The previous implementation of the word grid was messy and duplicative. There were multiple representations of the same conceptual object (char[][] wordSearchGrid and int[][] selectionGrid), as well as poor separation of UI and backend logic. To address this, the word grid is now represented by a custom class WordGrid and all non-UI grid functionality has been moved there.

The grid is now represented by a single 2D array of WordCells that hold all information needed to interact with the grid contents. WordCells contain the text content (cellContent), selection status (isSelected), as well as the new attributes isPartOfWord, isStartingCharacter, and parentWord.

When a word is placed, all WordCells belonging to it are marked as part of a word and their parent word is recorded. During grid generation, the parent word and it's child cells are added to the wordPositionMap hashmap. This allowed for checkForWord to be vastly simplified, as it doesn't have to randomly try every directional possibility upon cell selection. Now it simply checks if the cell has a true isPartOfWord and isStartingCharacter. This also allows for automatic highlighting of all associated cells on clicking a cell with a true isStartingCharacter value.