WIP Selection Fixes
Word selection is no longer an automatic toggle, instead the initial status selection (T/F) of the starting cell is recorded and applied to the whole word. This makes it so that unselected a cell that belongs to a selected word wont break autoselection. This helps with intersecting words, but they are still broken.